# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: NH Advertisement Compliance API
  description: Source contract for the advertisement-compliance PoC. Capability-specific paths and
    schemas are added only when their implementation slice begins.
  version: 0.8.0
servers:
  - url: /api/v1
...
paths:
...
  /advertisements/{advertisementId}/reviews:
    post:
      tags:
        - Reviews
      operationId: requestAdvertisementReview
      description: Persists review/job/steps before minimal Redis delivery.
      parameters:
        - $ref: "#/components/parameters/AdvertisementId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "202":
          description: Review job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
      security:
        - BearerAuth: []
    get:
      tags:
        - Reviews
      operationId: listAdvertisementReviews
      description: Lists immutable review rounds for an advertisement.
      parameters:
        - $ref: "#/components/parameters/AdvertisementId"
      responses:
        "200":
          description: Immutable review history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewHistory"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
  /reviews/{reviewId}/status:
    get:
      tags:
        - Reviews
      operationId: getReviewStatus
      description: Returns PostgreSQL-sourced job and step progress.
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      responses:
        "200":
          description: PostgreSQL-sourced job and step progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewProgress"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
  /reviews/{reviewId}/events:
    get:
      tags:
        - Reviews
      operationId: streamReviewProgress
      description: Streams PostgreSQL-sourced review progress changes as server-sent events.
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      responses:
        "200":
          description: Progress events. Each `progress` event contains a ReviewProgress payload.
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
  /reviews/{reviewId}/rerun:
    post:
      tags:
        - Reviews
      operationId: rerunReview
      description: Creates a new review round without overwriting history.
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RerunReviewRequest"
      responses:
        "202":
          description: New review round accepted without overwriting history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RerunReviewAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
      security:
        - BearerAuth: []
  /reviews/{reviewId}/summary:
    get:
      tags:
        - Review Results
      operationId: getReviewSummary
      description: Returns the immutable standard-version snapshot and risk summary.
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      responses:
        "200":
          description: Scoped review result returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
  /reviews/{reviewId}/items:
    get:
      tags:
        - Review Results
      operationId: listReviewItems
      description: Lists scoped result items without hiding evidence insufficiency or search failure.
      parameters:
        - $ref: "#/components/parameters/ReviewId"
        - $ref: "#/components/parameters/ReviewTypeQuery"
        - $ref: "#/components/parameters/RiskLevelQuery"
        - $ref: "#/components/parameters/ResultStatusQuery"
        - $ref: "#/components/parameters/IncludeAppropriateQuery"
        - $ref: "#/components/parameters/EvidenceRequiredQuery"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Scoped review result returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItemPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
  /reviews/{reviewId}/items/{reviewItemId}:
    get:
      tags:
        - Review Results
      operationId: getReviewItem
      description: Returns risk rationale, frozen evidence versions, and optional display location.
      parameters:
        - $ref: "#/components/parameters/ReviewId"
        - $ref: "#/components/parameters/ReviewItemId"
      responses:
        "200":
          description: Scoped review result returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItemDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
  /reviews/{reviewId}/annotations:
    get:
      tags:
        - Annotations
      operationId: listReviewAnnotations
      description: Returns BOX, TEXT_HIGHLIGHT, or list fallback annotations.
      parameters:
        - $ref: "#/components/parameters/ReviewId"
        - $ref: "#/components/parameters/PageNoQuery"
        - $ref: "#/components/parameters/ReviewTypeQuery"
        - $ref: "#/components/parameters/RiskLevelQuery"
      responses:
        "200":
          description: Scoped review result returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnotationCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    CreateReviewRequest:
      type: object
      additionalProperties: false
      properties:
        standardEffectiveDate:
          type:
            - string
            - "null"
          format: date
        reviewTypes:
          type:
            - array
            - "null"
          items:
            $ref: "#/components/schemas/ReviewType"
          minItems: 1
          uniqueItems: true
        includeSuggestion:
          type: boolean
          default: true
        includeOpinionDraft:
          type: boolean
          default: false
        requestMemo:
          type:
            - string
            - "null"
          maxLength: 2000
...
    ErrorResponse:
      additionalProperties: false
      example:
        code: FILE_NOT_SUPPORTED
        details:
          - field: advertisementFile
            reason: jpg, jpeg, png, pdf, hwp, hwpx 파일만 업로드할 수 있습니다.
        message: 지원하지 않는 파일 형식입니다.
        timestamp: 2026-07-14T06:00:00Z
        traceId: req-example-000001
      properties:
        code:
          minLength: 1
          type: string
        details:
          items:
            additionalProperties: true
            type: object
          type: array
        message:
          minLength: 1
          type: string
        timestamp:
          format: date-time
          type: string
        traceId:
          minLength: 1
          type: string
      required:
        - code
        - message
        - traceId
        - timestamp
      type: object
...
    RerunReviewRequest:
      type: object
      additionalProperties: false
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 2000
        reviewTypes:
          type:
            - array
            - "null"
          items:
            $ref: "#/components/schemas/ReviewType"
          minItems: 1
          uniqueItems: true
...
    AIReviewStatus:
      type: string
      enum:
        - ANALYSIS_REQUESTED
        - ANALYZING
        - CHECK_REQUIRED
        - REVIEW_COMPLETED
        - REVIEW_FAILED
    ReviewJobStatus:
      type: string
      enum:
        - PENDING
        - RUNNING
        - RETRY_PENDING
        - STALE
        - COMPLETED
        - FAILED
        - FAILED_FINAL
        - CANCELED
    ReviewType:
      type: string
      enum:
        - REQUIRED_PHRASE
        - INTEREST_RATE
        - MISLEADING_EXPRESSION
        - PRODUCT_CONSISTENCY
        - VISIBILITY
        - OCR_QUALITY
    ReviewAccepted:
      type: object
      additionalProperties: false
      required:
        - reviewId
        - advertisementId
        - reviewStatus
        - jobId
        - standardEffectiveDate
        - standardVersionIds
        - requestedAt
      properties:
        reviewId:
          type: string
        advertisementId:
          type: string
        reviewStatus:
          const: ANALYSIS_REQUESTED
          type: string
        jobId:
          type: string
        standardEffectiveDate:
          type: string
          format: date
        standardVersionIds:
          type: array
          items:
            type: string
        requestedAt:
          type: string
          format: date-time
    ReviewStepStatus:
      type: object
      additionalProperties: false
      required:
        - stepCode
        - stepName
        - status
        - timeoutAt
      properties:
        stepCode:
          type: string
        stepName:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - RUNNING
            - RETRY_PENDING
            - COMPLETED
            - FAILED
            - SKIPPED
        timeoutAt:
          type:
            - string
            - "null"
          format: date-time
        failedReasonCode:
          type:
            - string
            - "null"
    ReviewProgress:
      type: object
      additionalProperties: false
      required:
        - reviewId
        - advertisementId
        - reviewStatus
        - jobId
        - jobStatus
        - progressRate
        - retryCount
        - maxRetries
        - isRetryable
        - timeoutAt
        - steps
        - updatedAt
      properties:
        reviewId:
          type: string
        advertisementId:
          type: string
        reviewStatus:
          $ref: "#/components/schemas/AIReviewStatus"
        jobId:
          type: string
        jobStatus:
          $ref: "#/components/schemas/ReviewJobStatus"
        currentStep:
          type:
            - string
            - "null"
        progressRate:
          type: number
          minimum: 0
          maximum: 100
        retryCount:
          type: integer
          minimum: 0
          maximum: 3
        maxRetries:
          const: 3
          type: integer
        nextRetryAt:
          type:
            - string
            - "null"
          format: date-time
        isRetryable:
          type: boolean
        failedReasonCode:
          type:
            - string
            - "null"
        failedReason:
          type:
            - string
            - "null"
        timeoutAt:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ReviewStepStatus"
        updatedAt:
          type: string
          format: date-time
    ReviewHistory:
      type: object
      additionalProperties: false
      required:
        - reviewId
        - reviewRound
        - reviewStatus
        - requestedAt
      properties:
        reviewId:
          type: string
        reviewRound:
          type: integer
          minimum: 1
        reviewStatus:
          $ref: "#/components/schemas/AIReviewStatus"
        overallRiskLevel:
          type:
            - string
            - "null"
        requestedAt:
          type: string
          format: date-time
        completedAt:
          type:
            - string
            - "null"
          format: date-time
    RerunReviewAccepted:
      type: object
      additionalProperties: false
      required:
        - newReviewId
        - previousReviewId
        - reviewStatus
        - jobId
      properties:
        newReviewId:
          type: string
        previousReviewId:
          type: string
        reviewStatus:
          const: ANALYSIS_REQUESTED
          type: string
        jobId:
          type: string
...
    RiskLevel:
      type: string
      enum:
        - HIGH
        - MEDIUM
        - LOW
        - CHECK_REQUIRED
    ReviewResultStatus:
      type: string
      enum:
        - APPROPRIATE
        - NEEDS_REVISION
        - NEEDS_CONFIRMATION
    EvidenceStatus:
      type: string
      enum:
        - CONNECTED
        - NOT_REQUIRED
        - INSUFFICIENT
        - SEARCH_UNAVAILABLE
...
    RiskRationale:
      type: object
      additionalProperties: false
      required:
        - riskLevel
        - policyVersion
        - reasonCodes
        - scoreDetail
      properties:
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        policyVersion:
          type: string
          minLength: 1
        reasonCodes:
          type: array
          items:
            type: string
            minLength: 1
          minItems: 1
          uniqueItems: true
        scoreDetail:
          $ref: "#/components/schemas/RiskScoreDetail"
    ReviewTypeSummary:
      type: object
      additionalProperties: false
      required:
        - reviewType
        - totalCount
        - needsRevisionCount
        - needsConfirmationCount
      properties:
        reviewType:
          $ref: "#/components/schemas/ReviewType"
        totalCount:
          type: integer
          minimum: 0
        needsRevisionCount:
          type: integer
          minimum: 0
        needsConfirmationCount:
          type: integer
          minimum: 0
    TopRisk:
      type: object
      additionalProperties: false
      required:
        - reviewItemId
        - riskLevel
        - riskPolicyVersion
        - riskReasonCodes
        - targetText
        - reason
        - evidenceStatus
      properties:
        reviewItemId:
          type: string
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        riskPolicyVersion:
          type: string
        riskReasonCodes:
          type: array
          items:
            type: string
        targetText:
          type: string
        reason:
          type: string
        evidenceStatus:
          $ref: "#/components/schemas/EvidenceStatus"
    ReviewSummary:
      type: object
      additionalProperties: false
      required:
        - reviewId
        - advertisementId
        - standardEffectiveDate
        - standardVersionIds
        - overallRiskLevel
        - totalItemCount
        - needsRevisionCount
        - needsConfirmationCount
        - reviewTypeSummary
        - topRisks
        - completedAt
      properties:
        reviewId:
          type: string
        advertisementId:
          type: string
        standardEffectiveDate:
          type: string
          format: date
        standardVersionIds:
          type: array
          items:
            type: string
        overallRiskLevel:
          $ref: "#/components/schemas/RiskLevel"
        totalItemCount:
          type: integer
          minimum: 0
        needsRevisionCount:
          type: integer
          minimum: 0
        needsConfirmationCount:
          type: integer
          minimum: 0
        reviewTypeSummary:
          type: array
          items:
            $ref: "#/components/schemas/ReviewTypeSummary"
        topRisks:
          type: array
          items:
            $ref: "#/components/schemas/TopRisk"
        completedAt:
          type: string
          format: date-time
    ReviewItemSummary:
      type: object
      additionalProperties: false
      required:
        - reviewItemId
        - reviewType
        - targetText
        - resultStatus
        - riskLevel
        - riskPolicyVersion
        - riskReasonCodes
        - reason
        - evidenceStatus
        - evidenceFailureCode
        - evidenceCount
        - pageNo
        - hasAnnotation
        - sourceEngine
        - sourceVersion
      properties:
        reviewItemId:
          type: string
        reviewType:
          $ref: "#/components/schemas/ReviewType"
        targetText:
          type: string
        resultStatus:
          $ref: "#/components/schemas/ReviewResultStatus"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        riskPolicyVersion:
          type: string
        riskReasonCodes:
          type: array
          items:
            type: string
        reason:
          type: string
        evidenceStatus:
          $ref: "#/components/schemas/EvidenceStatus"
        evidenceFailureCode:
          oneOf:
            - type: string
              enum:
                - RAG_SEARCH_UNAVAILABLE
                - RAG_SEARCH_FAILED
            - type: "null"
        evidenceCount:
          type: integer
          minimum: 0
          maximum: 5
        pageNo:
          type:
            - integer
            - "null"
          minimum: 1
        hasAnnotation:
          type: boolean
        sourceEngine:
          $ref: "#/components/schemas/ReviewEngineType"
        sourceVersion:
          type: string
          minLength: 1
    ReviewItemPage:
      type: object
      additionalProperties: false
      required:
        - contents
        - page
        - size
        - totalElements
        - totalPages
      properties:
        contents:
          type: array
          items:
            $ref: "#/components/schemas/ReviewItemSummary"
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    AnnotationCollection:
      type: object
      additionalProperties: false
      required:
        - reviewId
        - fileId
        - fileType
        - pageNo
        - annotations
      properties:
        reviewId:
          type: string
        fileId:
          type: string
        fileType:
          type: string
        pageNo:
          type:
            - integer
            - "null"
          minimum: 1
        annotations:
          type: array
          items:
            $ref: "#/components/schemas/Annotation"
    ReviewItemDetail:
      allOf:
        - $ref: "#/components/schemas/ReviewItemSummary"
        - type: object
          additionalProperties: false
          required:
            - riskRationale
            - evidences
            - recommendation
            - annotation
          properties:
            riskRationale:
              $ref: "#/components/schemas/RiskRationale"
            evidences:
              type: array
              maxItems: 3
              items:
                $ref: "#/components/schemas/ReviewItemEvidence"
            recommendation:
              type:
                - string
                - "null"
            annotation:
              oneOf:
                - $ref: "#/components/schemas/Annotation"
                - type: "null"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  parameters:
...
    Page:
      in: query
      name: page
      schema:
        default: 1
        minimum: 1
        type: integer
...
    Size:
      in: query
      name: size
      schema:
        default: 20
        maximum: 100
        minimum: 1
        type: integer
...
    AdvertisementId:
      in: path
      name: advertisementId
      required: true
      schema:
        type: string
        pattern: ^ADV-[A-Za-z0-9-]+$
    ReviewId:
      in: path
      name: reviewId
      required: true
      schema:
        type: string
        pattern: ^REV-[A-Za-z0-9-]+$
    ReviewItemId:
      in: path
      name: reviewItemId
      required: true
      schema:
        type: string
        pattern: ^ITEM-[A-Za-z0-9-]+$
    ReviewTypeQuery:
      in: query
      name: reviewType
      schema:
        $ref: "#/components/schemas/ReviewType"
    RiskLevelQuery:
      in: query
      name: riskLevel
      schema:
        $ref: "#/components/schemas/RiskLevel"
    ResultStatusQuery:
      in: query
      name: resultStatus
      schema:
        $ref: "#/components/schemas/ReviewResultStatus"
    EvidenceRequiredQuery:
      in: query
      name: evidenceRequired
      schema:
        type: boolean
    IncludeAppropriateQuery:
      in: query
      name: includeAppropriate
      description: Includes appropriate items; omitted values show actionable results only.
      schema:
        type: boolean
        default: false
    PageNoQuery:
      in: query
      name: pageNo
      schema:
        type: integer
        minimum: 1
...
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: Request validation or file integrity failed.
    Conflict:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: Duplicate or conflicting request.
    Forbidden:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: Authenticated caller is outside the permitted role or resource scope.
    NotFound:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: Resource not found.
...
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: Authentication failed without revealing account state.
...